feat(api): Add back compat  shim for new protocol API
