+++ api/opentrons/protocol_api/__init__.py
@@ -9,18 +9,17 @@ import os
 from typing import Callable, List, Dict
 
 from opentrons.protocol_api.labware import Well, Labware, load
+from . import back_compat
 
-
-def run(protocol_callable: Callable[['ProtocolContext'], None] = None,
+def run(protocol_bytes: bytes = None,
         protocol_json: str = None,
         simulate: bool = False,
         context: 'ProtocolContext' = None):
         """ Create a ProtocolRunner instance from one of a variety of protocol
         sources.
 
-        :param protocol_callable: If the protocol can be represented as a
-        Python callable, pass it here. The callable should take a context
-        as its only argument.
+        :param protocol_bytes: If the protocol is a Python protocol, pass the
+        file contents here.
         :param protocol_json: If the protocol is a json file, pass the contents
         here.
         :param simulate: True to simulate; False to execute. If thsi is not an
@@ -30,10 +29,10 @@ def run(protocol_callable: Callable[['ProtocolContext'], None] = None,
         """
         if not os.environ.get('RUNNING_ON_PI'):
             simulate = True # noqa - will be used later
-        if None is context:
+        if None is context and simulate:
             context = ProtocolContext()
-        if protocol_callable:
-            pass
+        if protocol_bytes:
+            back_compat.run(protocol_bytes, context)
         elif protocol_json:
             pass
 
@@ -50,7 +49,8 @@ class ProtocolContext:
         pass
 
     def load_labware(
-            self, labware_obj: Labware, location: str):
+            self, labware_obj: Labware, location: str,
+            label: str=None, share: bool = False):
         """ Specify the presence of a piece of labware on the OT2 deck.
 
         This function loads the labware specified by ``labware``
